Add mk_ty and mk_substs, move InternAs
